requirements.txt
alabaster==1.0.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
authlib==1.6.3
babel==2.17.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.6
cryptography==45.0.7
decorator==5.2.1
django==5.2.6
django-cors-headers==4.8.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-stubs==5.2.5
django-stubs-ext==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djangorestframework-stubs==3.16.2
docutils==0.21.2
dparse==0.6.4
drf-spectacular==0.28.0
executing==2.2.1
factory-boy==3.3.3
faker==37.6.0
filelock==3.19.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.1.0
ipython==9.5.0
ipython-pygments-lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==4.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-extensions==1.1.0
nltk==3.9.1
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pillow==11.3.0
pip==25.2
platformdirs==4.4.0
pluggy==1.6.0
promise==2.3
prompt-toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.3
pycparser==2.23
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.19.2
pyjwt==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pyyaml==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
roman-numerals-py==3.1.0
rpds-py==0.27.1
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.12
ruff==0.13.0
safety==3.6.1
safety-schemas==0.0.14
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
stack-data==0.6.3
structlog==25.4.0
tenacity==9.1.2
text-unidecode==1.3
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.4
types-pyyaml==6.0.12.20250822
types-requests==2.32.4.20250913
typing-extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.17
vine==5.1.0
wcwidth==0.2.13
